+++ queries-core/src/pipeline/mod.rs
@@ -3,4 +3,4 @@ pub mod intent_classifier;
 pub mod tokens_classifier;
 pub mod slot_filler;
 
-pub type Probability = f64;
+pub type Probability = f32;
